+++ shellfoundry/bootstrap.py
@@ -1,11 +1,18 @@
+import os
+
 import click
 import pkg_resources
+from cloudshell.rest.api import PackagingRestApiClient
+from cloudshell.rest.exceptions import ShellNotFoundException
+from requests import post
 
 from shellfoundry.commands.dist_command import DistCommandExecutor
+from shellfoundry.commands.generate_command import GenerateCommandExecutor
 from shellfoundry.commands.install_command import InstallCommandExecutor
 from shellfoundry.commands.list_command import ListCommandExecutor
 from shellfoundry.commands.new_command import NewCommandExecutor
 from shellfoundry.commands.pack_command import PackCommandExecutor
+from shellfoundry.utilities.shell_package import ShellPackage
 
 
 @click.group()
@@ -65,3 +72,12 @@ def dist():
     """
     PackCommandExecutor().pack()
     DistCommandExecutor().dist()
+
+
+@cli.command()
+def generate():
+    """
+    Generates Python driver data model to be used in driver code
+    """
+    PackCommandExecutor().pack()
+    GenerateCommandExecutor().generate()